add NoAccessModal.js and put in DataSFilter.js and AddFilter.js
